backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
dj-database-url==1.2.0
dj-rest-auth==3.0.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.3.1
inflection==0.5.1
jsonschema==4.17.3
psycopg2-binary==2.9.5
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
